livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "livekit-agents[openai]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.2",
]
